* Made all disassemblers return MachineInstruction to make more generic operation possible. * IProcessorArchitecture.CreateDisassembler takes a single imagereader parameter now. * ImageReader supports big-endian operation. * Fixed bugs in BinHexDecoder; it is now code-complete + M68k - Motorola 68000+ support.

+++ trunk/src/Arch/Intel/LongAddRewriter.cs
@@ -66,7 +66,7 @@ namespace Decompiler.Arch.Intel
 			{
 				if (instrs[i].code == next)
 					return i;
-				if ((instrs[i].DefCc() & FlagM.CF) != 0)
+				if ((instrs[i].DefCc() & (uint) FlagM.CF) != 0)
 					break;
 			}
 			return -1;
